[AIRFLOW-2110] Moved tests from core to http hook and added new test case
Added test case test_header_from_extra_and_run_method_are_merged
